Fix: Correct uptime accuracy calculation for undeployed devices
