+++ src/device-registry/utils/common/index.js
@@ -21,6 +21,8 @@ const generateFilter = require("./generate-filter");
 const handleResponse = require("./responseHandler");
 const distance = require("./distance");
 const translate = require("./translate");
+const { getUptimeAccuracyUpdateObject } = require("./uptime.util");
+
 const claimTokenUtil = require("./claimToken.util");
 const { deduplicator } = require("./slack-dedup-utility");
 const ActivityLogger = require("./activity-logger.util");
@@ -28,6 +30,7 @@ const ActivityLogger = require("./activity-logger.util");
 module.exports = {
   claimTokenUtil,
   ActivityLogger,
+  getUptimeAccuracyUpdateObject,
   deduplicator,
   translate,
   distance,
